python/requirements.txt
altgraph==0.17.4
asgiref==3.7.2
Django==4.2.6
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
packaging==23.2
pefile==2023.2.7
psycopg2-binary==2.9.9
pyinstaller==6.3.0
pyinstaller-hooks-contrib==2024.0
PyJWT==2.8.0
python-dotenv==1.0.0
pytz==2024.1
pywin32-ctypes==0.2.2
setuptools==69.0.3
sqlparse==0.4.4
tzdata==2023.4